requirements.txt
streamlit
matplotlib
nltk
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose[cryptography]